# [REVIEW] Optimize `DataFrame` creation across code-base