--- a/python/cudf/cudf/core/column/methods.py
+++ b/python/cudf/cudf/core/column/methods.py
@@ -84,7 +84,7 @@ class ColumnMethods:
                     idx.names = None
                     return idx
                 else:
-                    return self._parent._constructor_expanddim(
+                    return self._parent._constructor_expanddim._from_data(
                         data=table._data, index=self._parent.index
                     )
             elif isinstance(self._parent, cudf.Series):