--- a/python/cudf/cudf/tests/test_pack.py
+++ b/python/cudf/cudf/tests/test_pack.py
@@ -62,7 +62,7 @@ def assert_packed_frame_equality(df):
     packed = pack(df)
     del df
     tbl = unpack(packed)
-    unpacked = DataFrame(tbl._data, tbl._index)
+    unpacked = DataFrame._from_data(tbl._data, tbl._index)
 
     assert_eq(unpacked, pdf)
 
@@ -198,14 +198,14 @@ def check_packed_pickled_equality(df):
         for b in buffers:
             assert isinstance(b, pickle.PickleBuffer)
         tbl = unpack(pickle.loads(serialbytes, buffers=buffers))
-        loaded = DataFrame(tbl._data, tbl._index)
+        loaded = DataFrame._from_data(tbl._data, tbl._index)
         assert_eq(loaded, df)
 
 
 def assert_packed_frame_picklable(df):
     serialbytes = pickle.dumps(pack(df))
     tbl = unpack(pickle.loads(serialbytes))
-    loaded = DataFrame(tbl._data, tbl._index)
+    loaded = DataFrame._from_data(tbl._data, tbl._index)
     assert_eq(loaded, df)
 
 
@@ -271,7 +271,7 @@ def assert_packed_frame_serializable(df):
     packed = pack(df)
     header, frames = packed.serialize()
     tbl = unpack(packed.deserialize(header, frames))
-    loaded = DataFrame(tbl._data, tbl._index)
+    loaded = DataFrame._from_data(tbl._data, tbl._index)
     assert_eq(loaded, df)