After internal profiling, it appears that `_init_from_dict_like` was being repeatedly called even for `ColumnAccessor` inputs, which is not necessary and is an expensive method where we do proper index re-alignment. This PR handles by raising an error when a developer tries to create a `DataFrame` with `ColumnAccessor`, and also changes multiple of those instances where such calls are present to `_from_data` call which will now avoid going through `_init_from_dict_like` method for `ColumnAccessor` input. For a dataframe of shape `30_00_000 x 3` the speed up is about **2.5x to 4x**.

```python
In [1]: import cudf

In [2]: x = cudf.DataFrame(
   ...:     {
   ...:         "a": [1, 2, 3] * 10_00_000,
   ...:         "b": [3, 4, 5] * 10_00_000,
   ...:         "c": ["a", "b", "c"] * 10_00_000,
   ...:     },
   ...:     index=list(range(0, 3000000)),
   ...: )

In [3]: x
Out[3]: 
         a  b  c
0        1  3  a
1        2  4  b
2        3  5  c
3        1  3  a
4        2  4  b
...     .. .. ..
2999995  2  4  b
2999996  3  5  c
2999997  1  3  a
2999998  2  4  b
2999999  3  5  c

[3000000 rows x 3 columns]

In [4]: x.index
Out[4]: 
Int64Index([      0,       1,       2,       3,       4,       5,       6,
                  7,       8,       9,
            ...
            2999990, 2999991, 2999992, 2999993, 2999994, 2999995, 2999996,
            2999997, 2999998, 2999999],
           dtype='int64', length=3000000)
```
On `branch-22.04`:
```python
In [3]: %timeit cudf.DataFrame(x._data, x._index)
35.4 µs ± 240 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [4]: %timeit cudf.DataFrame(x._data)
40.6 µs ± 242 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

```

This `PR`:

```python
In [4]: %timeit cudf.DataFrame._from_data(x._data, x._index)
8.11 µs ± 169 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [3]: %timeit cudf.DataFrame._from_data(x._data)
16.9 µs ± 220 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```





<!--

Thank you for contributing to cuDF :)

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are being
   made.

2. Please ensure that you have written units tests for the changes made/features
   added.

3. There are CI checks in place to enforce that committed code follows our style
   and syntax standards. Please see our contribution guide in `CONTRIBUTING.MD`
   in the project root for more information about the checks we perform and how
   you can run them locally.

4. If you are closing an issue please use one of the automatic closing words as
   noted here: https://help.github.com/articles/closing-issues-using-keywords/

5. If your pull request is not ready for review but you want to make use of the
   continuous integration testing facilities please mark your pull request as Draft.
   https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/changing-the-stage-of-a-pull-request#converting-a-pull-request-to-a-draft

6. If your pull request is ready to be reviewed without requiring additional
   work on top of it, then remove it from "Draft" and make it "Ready for Review".
   https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/changing-the-stage-of-a-pull-request#marking-a-pull-request-as-ready-for-review

   If assistance is required to complete the functionality, for example when the
   C/C++ code of a feature is complete but Python bindings are still required,
   then add the label `help wanted` so that others can triage and assist.
   The additional changes then can be implemented on top of the same PR.
   If the assistance is done by members of the rapidsAI team, then no
   additional actions are required by the creator of the original PR for this,
   otherwise the original author of the PR needs to give permission to the
   person(s) assisting to commit to their personal fork of the project. If that
   doesn't happen then a new PR based on the code of the original PR can be
   opened by the person assisting, which then will be the PR that will be
   merged.

7. Once all work has been done and review has taken place please do not add
   features or make changes out of the scope of those requested by the reviewer
   (doing this just add delays as already reviewed code ends up having to be
   re-reviewed/it is hard to tell what is new etc!). Further, please do not
   rebase your branch on the target branch, force push, or rewrite history.
   Doing any of these causes the context of any comments made by reviewers to be lost.
   If conflicts occur against the target branch they should be resolved by
   merging the target branch into the branch used for making the pull request.

Many thanks in advance for your cooperation!

-->