@vyasr 2022-02-07
To be clear, there are _a ton_ of places in cudf that are doing inefficient things right now when it comes to materializing more intermediates than we need. We need to do a comprehensive review to remove things like this.