@galipremsagar 2022-02-07
After an offline discussion with @vyasr we decided to raise an error if someone passes a `ColumnAccessor` to `DataFrame/Series` constructors thus forcing devs to use a much faster `Series/DataFrame._from_data` method. I have also now changed the internal code that was using the ineffecient code path before.